@@ -4,31 +4,30 @@
   "license": "The Apache Software License, Version 2.0",
   "angular-cli": {},
   "scripts": {
+    "ng-high-memory": "node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng",
     "ng": "ng",
-    "start": "ng serve --prod",
+    "start": "ng serve",
     "start-with-proxy": "ng serve --proxy-config proxy.conf.json",
-    "publish-gh-pages": "rm -rf dist; ng build --output-path=dist --dev --base-href \"https://complex-portal.github.io/complex-portal-view/\"; sudo ngh",
-    "build-prod": "rm -rf dist; ng build --prod --base-href \"/complexportal/\"",
+    "publish-gh-pages": "rm -rf dist; node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --output-path=dist --prod --base-href \"https://complex-portal.github.io/complex-portal-view/\"; sudo ngh",
+    "build-prod": "rm -rf dist; node --max_old_space_size=8000 ./node_modules/@angular/cli/bin/ng build --prod --base-href \"/complexportal/\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.8",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.8",
-    "@angular/compiler": "^5.2.8",
-    "@angular/core": "^5.2.8",
-    "@angular/forms": "^5.2.8",
-    "@angular/http": "^5.2.8",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.8",
-    "@angular/platform-browser-dynamic": "^5.2.8",
-    "@angular/router": "^5.2.8",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ebi-gene-expression-group/anatomogram": "^2.4.0",
     "angular-cli-ghpages": "^0.5.2",
-    "angulartics2": "^2.4.0",
     "animate.css": "^3.5.2",
     "classlist.js": "^1.1.20150312",
     "complexviewer": "^2.1.14",
@@ -37,25 +36,30 @@
     "expose-loader": "^0.7.5",
     "jquery": "^3.4.1",
     "litemol": "https://github.com/dsehnal/LiteMol/tarball/master",
-    "ngx-markdown": "^1.5.2",
-    "ngx-page-scroll": "^4.0.2",
-    "ngx-toastr": "^8.3.0",
+    "ngx-google-analytics": "^8.1.1",
+    "ngx-markdown": "^8.2.2",
+    "ngx-page-scroll": "^6.0.2",
+    "ngx-page-scroll-core": "^6.0.2",
+    "ngx-toastr": "^10.2.0",
     "npm": "^5.7.1",
     "react": "^15.6.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.6.7",
+    "rxjs-compat": "^6.6.7",
     "ts-helpers": "^1.1.2",
     "ts-md5": "^1.2.4",
-    "zone.js": "^0.8.20"
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^6.2.9",
-    "@angular/compiler-cli": "^5.2.8",
-    "@angular/language-service": "^5.2.8",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/foundation": "~5.5.30",
     "@types/jasmine": "2.8.6",
     "@types/jquery": "~3.3.36",
     "@types/node": "~14.0.1",
-    "codelyzer": "^4.2.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "^3.1.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^2.0.0",
@@ -67,6 +71,6 @@
     "protractor": "^5.3.0",
     "ts-node": "^5.0.1",
     "tslint": "^5.9.1",
-    "typescript": "~2.9.2"
+    "typescript": "~3.5.3"
   }
 }
