@@ -7,7 +7,6 @@ import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app-routing.module';
 import {ComplexPortalService} from './complex/shared/service/complex-portal.service';
 import {NotificationService} from './shared/notification/service/notification.service';
-import {Angulartics2GoogleAnalytics, Angulartics2Module} from 'angulartics2';
 import {ReactomeService} from './complex/complex-details/complex-function/reactome-crossreferences/service/reactome.service';
 import {EuroPmcService} from './complex/complex-details/complex-references/euro-pmc-crossreferences/service/euro-pmc.service';
 import {OlsService} from './shared/ols/service/ols.service';
@@ -17,17 +16,20 @@ import {ProgressBarComponent} from './shared/loading-indicators/progress-bar/pro
 import {ToastrModule} from 'ngx-toastr';
 import {LocalSearchComponent} from './search/local-search/local-search.component';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
-import {MatProgressBarModule} from '@angular/material';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {OntologiesComponent} from './ontologies/ontologies.component';
-import {GoogleAnalyticsService} from './shared/google-analytics/service/google-analytics.service';
+import {AnalyticsService} from './shared/google-analytics/service/analytics.service';
 import {LocalStorageService} from './shared/local-storage/service/local-storage.service';
 import {SearchService} from './search/service/search.service';
 import {MarkdownModule} from 'ngx-markdown';
 import {HomeModule} from './home/home.module';
-import {Router, RouterModule} from '@angular/router';
+import {RouterModule} from '@angular/router';
 import {AboutComponent} from './navigation/about/about.component';
 import {DownloadComponent} from './download/download.component';
 import {DocumentationComponent} from './navigation/documentation/documentation.component';
+import {NgxPageScrollCoreModule} from 'ngx-page-scroll-core';
+import {NgxGoogleAnalyticsModule, NgxGoogleAnalyticsRouterModule} from 'ngx-google-analytics';
+import {environment} from '../environments/environment';
 
 @NgModule({
   declarations: [
@@ -46,9 +48,11 @@ import {DocumentationComponent} from './navigation/documentation/documentation.c
     HttpClientModule,
     NoopAnimationsModule,
     ToastrModule.forRoot(),
-    Angulartics2Module.forRoot([Angulartics2GoogleAnalytics]),
+    NgxPageScrollCoreModule.forRoot({scrollOffset: 50}),
     MatProgressBarModule,
     MarkdownModule.forRoot(),
+    NgxGoogleAnalyticsModule.forRoot(environment.analytics_id),
+    NgxGoogleAnalyticsRouterModule,
     HomeModule,
     AppRoutingModule,
   ],
@@ -61,7 +65,7 @@ import {DocumentationComponent} from './navigation/documentation/documentation.c
     BasketService,
     LocalStorageService,
     Md5,
-    GoogleAnalyticsService,
+    AnalyticsService,
     SearchService
   ],
   bootstrap: [AppComponent]
