@@ -1,9 +1,11 @@
+import {catchError} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpErrorResponse, HttpResponse} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import {HttpClient, HttpErrorResponse} from '@angular/common/http';
+
+
 import {environment} from '../../../../../../environments/environment';
+import {Observable} from 'rxjs/Observable';
+import {throwError} from 'rxjs/internal/observable/throwError';
 
 const baseURL = environment.reactome_base_url;
 
@@ -18,9 +20,9 @@ export class ReactomeService {
    * @param id - a reactome stable identifier
    * @returns {Observable<R|T>}
    */
-  public findRelatedPathways(id: string) {
-    return this.http.get(baseURL + '/ContentService/data/pathways/low/entity/' + id)
-      .catch(this.handleError);
+  public findRelatedPathways(id: string): Observable<Pathway[]> {
+    return this.http.get<Pathway[]>(baseURL + '/ContentService/data/pathways/low/entity/' + id)
+      .pipe(catchError(this.handleError));
   }
 
   /**
@@ -28,17 +30,32 @@ export class ReactomeService {
    * @param id - a complex stable identifier
    * @returns {Observable<R|T>}
    */
-  public getComplexName(id: string) {
-    return this.http.get(baseURL + '/ContentService/data/query/' + id + '/displayName', {observe: 'response', responseType: 'text'})
-      // .map((res: HttpResponse) => res.body.text())
-      .catch(this.handleError);
+  public getComplexName(id: string): Observable<string> {
+    return this.http.get(baseURL + '/ContentService/data/query/' + id + '/displayName', {observe: 'body', responseType: 'text'})
+      .pipe(catchError(this.handleError));
   }
 
   private handleError(err: HttpErrorResponse | any): Observable<any> {
     if (err.error instanceof Error) {
-      return Observable.throw(err);
+      return throwError(err);
     } else {
       console.error(err.message ? err.message : err.toString());
     }
   }
+}
+
+export interface Pathway {
+  dbId: number;
+  displayName: string;
+  stId: string;
+  stIdVersion: string;
+  isInDisease: boolean;
+  isInferred: boolean;
+  name: string[];
+  releaseDate: string;
+  speciesName: string;
+  hasDiagram: boolean;
+  hasEHLD: boolean;
+  schemaClass: string;
+  className: string;
 }
