@@ -1,9 +1,12 @@
+import {catchError} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpErrorResponse, HttpResponse} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import {HttpClient, HttpErrorResponse} from '@angular/common/http';
+
+
 import {environment} from '../../../../../../environments/environment';
+import {EuroPMCResponse} from '../model/EuroPMCResponse';
+import {throwError} from 'rxjs/internal/observable/throwError';
+import {Observable} from 'rxjs/Observable';
 
 const baseURL = environment.europepmc_base_url;
 
@@ -13,14 +16,14 @@ export class EuroPmcService {
   constructor(private http: HttpClient) {
   }
 
-  getPublicationInformation(id: string) {
-    return this.http.get(baseURL + '/webservices/rest/search?query=ext_id:' + id + '%20src:med&format=json')
-      .catch(this.handleError);
+  getPublicationInformation(id: string): Observable<EuroPMCResponse> {
+    return this.http.get<EuroPMCResponse>(baseURL + '/webservices/rest/search?query=ext_id:' + id + '%20src:med&format=json')
+      .pipe(catchError(this.handleError));
   }
 
   private handleError(err: HttpErrorResponse | any): Observable<any> {
     if (err.error instanceof Error) {
-      return Observable.throw(err);
+      return throwError(err);
     } else {
       console.error(err.message ? err.message : err.toString());
     }
