@@ -4,15 +4,20 @@ import {RouterModule} from '@angular/router';
 
 @NgModule({
   imports: [RouterModule.forChild([
-    {path: 'organisms', loadChildren: 'app/complex/complex-organisms/complex-organisms.module#ComplexOrganismsModule'},
-    {path: 'search', loadChildren: 'app/complex/complex-results/complex-results.module#ComplexResultsModule'},
-    {path: ':id', loadChildren: 'app/complex/complex-details/complex-details.module#ComplexDetailsModule'},
+    {
+      path: 'organisms',
+      loadChildren: () => import('app/complex/complex-organisms/complex-organisms.module').then(m => m.ComplexOrganismsModule)
+    },
+    {
+      path: 'search',
+      loadChildren: () => import('app/complex/complex-results/complex-results.module').then(m => m.ComplexResultsModule)
+    },
+    {
+      path: ':id',
+      loadChildren: () => import('app/complex/complex-details/complex-details.module').then(m => m.ComplexDetailsModule)
+    },
   ]),
     CommonModule,
-  ],
-  declarations: [
-  ],
-  exports: [
   ]
 })
 export class ComplexModule {
