@@ -1,10 +1,9 @@
-import {Injectable} from '@angular/core';
 import {HttpClient, HttpErrorResponse} from '@angular/common/http';
-
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import {Injectable} from '@angular/core';
 import {environment} from '../../../../environments/environment';
+import {catchError, map} from 'rxjs/operators';
+import {Observable} from 'rxjs/Observable';
+import {throwError} from 'rxjs/internal/observable/throwError';
 
 const baseURL = environment.ols_base_url;
 
@@ -21,9 +20,9 @@ export class OlsService {
    * @returns {Observable<R>}
    */
   getOrphaNetName(id: string) {
-    return this.http.get(baseURL + '/ordo/terms?iri=http://www.orpha.net/ORDO/' + id.replace(':', '_'))
-      .map((response: Response) => response)
-      .catch(this.handleError);
+    return this.http.get(baseURL + '/ordo/terms?iri=http://www.orpha.net/ORDO/' + id.replace(':', '_')).pipe(
+      map((response: Response) => response),
+      catchError(this.handleError));
   }
 
   /**
@@ -32,13 +31,13 @@ export class OlsService {
    * @returns {Observable<R>}
    */
   getEfoName(id: string) {
-    return this.http.get(baseURL + '/efo/terms?iri=http://www.ebi.ac.uk/efo/' + id.replace(':', '_'))
-      .catch(this.handleError);
+    return this.http.get(baseURL + '/efo/terms?iri=http://www.ebi.ac.uk/efo/' + id.replace(':', '_')).pipe(
+      catchError(this.handleError));
   }
 
   private handleError(err: HttpErrorResponse | any): Observable<any> {
     if (err.error instanceof Error) {
-      return Observable.throw(err);
+      return throwError(err);
     } else {
       console.error(err.message ? err.message : err.toString());
     }
